requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]
pydantic
pydantic[email]
email-validator
python-multipart
alembic
langchain
langchain-openai
openai
pgvector
tiktoken